requirements.txt

Flask==2.3.3
pyserial==3.5
pyocd==0.34.2
